merge: develop back to feature

Created-by: Fancy1234567
Commit-by: chensheng12345;xiaoyao42;ZYQ5333;kendel11
Merged-by: xiaowenzihwl
Description: <!--  Thanks for sending a pull request!  Here are some tips for you:

1) If this is your first time, please read our contributor guidelines: https://gitcode.com/openJiuwen/community/blob/master/CONTRIBUTING.md

2) If you want to contribute your code but don't know who will review and merge, please add label `openjiuwen-assistant` to the pull request, we will find and do it as soon as possible.
-->

**What type of PR is this?**
<!-- 
选择下面一种标签替换下方 `/kind <label>`，可选标签类型有：
- /kind bug 
- /kind task
- /kind feature
- /kind refactor
- /kind clean_code
如PR描述不符合规范，修改PR描述后需要/check-pr重新检查PR规范。
-->
/kind <label>


**Self-checklist**:（**请自检，在[ ]内打上x，我们将检视你的完成情况，否则会导致pr无法合入**）

+ - [ ] **设计**：PR对应的方案是否已经经过Maintainer评审，方案检视意见是否均已答复并完成方案修改
+ - [ ] **测试**：PR中的代码是否已有UT/ST测试用例进行充分的覆盖，新增测试用例是否随本PR一并上库或已经上库
+ - [ ] **验证**：PR描述信息中是否已包含对该PR对应的Feature、Refactor、Bugfix的预期目标达成情况的详细验证结果描述
+ - [ ] **接口**：是否涉及对外接口变更，相应变更已得到接口评审组织的通过，API对应的注释信息已经刷新正确
+ - [ ] **文档**：是否涉及官网文档修改，如果涉及请及时提交资料到Doc仓

<!-- **Special notes for your reviewers**: -->
<!-- + - [ ] 是否导致无法前向兼容 -->
<!-- + - [ ] 是否涉及依赖的三方库变更 -->


See merge request: openJiuwen/agent-studio!666

+++ backend/upgrade/sqlite/alembic_agent/versions/fix_sqlite_agent_update_kb_and_memory_8f4846812221.py
@@ -8,6 +8,7 @@ Create Date: 2026-02-07 11:50:21.774021
 from typing import Sequence, Union
 
 from alembic import op
+from openjiuwen_studio.core.database.migration_utils import table_exists, column_exists, index_exists
 import sqlalchemy as sa
 from sqlalchemy import inspect
 
@@ -17,30 +18,6 @@ revision: str = '8f4846812221'
 down_revision: Union[str, Sequence[str], None] = '07acd03b1173'
 branch_labels: Union[str, Sequence[str], None] = None
 depends_on: Union[str, Sequence[str], None] = None
-
-
-def table_exists(table_name):
-    bind = op.get_bind()
-    inspector = inspect(bind)
-    return inspector.has_table(table_name)
-
-
-def column_exists(table_name, column_name):
-    bind = op.get_bind()
-    inspector = inspect(bind)
-    if not inspector.has_table(table_name):
-        return False
-    columns = [c['name'] for c in inspector.get_columns(table_name)]
-    return column_name in columns
-
-
-def index_exists(table_name, index_name):
-    bind = op.get_bind()
-    inspector = inspect(bind)
-    if not inspector.has_table(table_name):
-        return False
-    indexes = [i['name'] for i in inspector.get_indexes(table_name)]
-    return index_name in indexes
 
 
 def upgrade() -> None:
@@ -140,5 +117,6 @@ def downgrade() -> None:
             if index_exists('memory_base', 'idx_space_id_memory_base'):
                 batch_op.drop_index('idx_space_id_memory_base')
 
+    if table_exists('memory_base'):
         op.drop_table('memory_base')
     # ### end Alembic commands ###
